fms: Mutate route to have one shareable reference
If the reference to the route is replaced, any reference made to this route
becomes invalid. Thus, we mutate the route inside a RefCell so we can hold
multiple references to the route while also mutating it within the cell.

Furthermore, we set the cruise level and speed on the route when decoding
it. This is still WIP since the values are not properly updated whenever a new
route is decoded with other cruise values.

Signed-off-by: Joe Pearson <pearjo@protonmail.com>