fp: Change Performance to be a struct
Changing the performance to be a struct rather than a trait makes working with
it easier. We can now store it in e.g. a planner struct without the need to
propagate the generic property all the way up to the FMS.

Creating the performance from a function provides us the same flexibility as we
had by implementing the performance on another type.

Signed-off-by: Joe Pearson <pearjo@protonmail.com>