c: Add script to build and format the efb.h header
This uses cbindgen to generate the header from the source files and clang-format
to do the formatting. The additional formatting step is needed since cbindgen
does not provide the desired options to format and we focus with this tool on
the process of creating the header in the first place.

Signed-off-by: Joe Pearson <pearjo@protonmail.com>