fms: Simplify passing around of Route
Rather than ref counting the Route with a RefCell to mutate it and borrow checks
during runtime, we make the Route "simple" field of the FMS. The need to have
the Route reference counted and mutable was to use it in other structs as
well. This need is eliminated by using the builder pattern for the FMS sub
systems. This way we have all our user data within the builder and build the
system once the route changes. Since we would have needed to reconstruct the
systems anyhow on every route change, we don't have a performance penalty from
this simplification.

On the C-API we simplify things too, by letting the EfbRoute just be a wrapper
around a heap pointer to the FMS route. This should help to implement e.g. an
iterator for the legs which we can use safely from C.

Signed-off-by: Joe Pearson <pearjo@protonmail.com>