fms: Own flight planning instead of builder
Owning the flight planning lets other refer to the planning easier by reference
rather than every caller get a new planning. This allows furthermore, that we
can have more than one builder and just use a builder to build the flight
planning of the FMS. This makes working with an FFI also easier, since the
foreign language can create a builder, populate it with all information and pass
it then on to the FMS to build the planning.

Signed-off-by: Joe Pearson <pearjo@protonmail.com>