fp: Move flight planning module
The concept of FMS sub-systems turned out to be not that helpful and was hard to
draw a line of what should be a sub-system and what a module on it's
own (e.g. aircraft vs. flight planning). Thus, the planning and it's builder are
moved to the `fp` module.

Signed-off-by: Joe Pearson <pearjo@protonmail.com>